[Chetana] Added: getallbook,getBookByID,deleteBook,UpdateBook API

+++ CommonLayer/BookModel/GetBookModel.cs
@@ -16,14 +16,13 @@ namespace CommonLayer.BookModel
 
         public int Quantity { get; set; }
 
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
-        public double DiscountPrice { get; set; }
+        public decimal DiscountPrice { get; set; }
 
-        public float TotalRating { get; set; }
+        public double TotalRating { get; set; }
 
         public int RatingCount { get; set; }
-
         public string BookImg { get; set; }
     }
 }
